Country id type changed to Short

+++ src/main/java/ru/skillbox/zerone/backend/model/dto/response/VKCountryResponseDTO.java
@@ -15,7 +15,7 @@ public class VKCountryResponseDTO {
 
     @Data
     public static class CountryDTO {
-      private long id;
+      private short id;
       private String title;
     }
   }
